requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
azure-core==1.32.0
azure-storage-blob==12.23.1
binho-host-adapter==0.1.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dht11==0.1.0
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
isodate==0.7.2
pycparser==2.22
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
